// vulpes_Front-End-adriana-dev2/src/Interfaces/anexos/anexo3.tsx
export interface IAnexo3 {
  id_plano: number;
  raca: string;
  grupo: string;
  classe: string;
  pastagem: string;
  espaco: string;
  superfice: string;
  forragem: string;
  alimentos: string;
  outras: string;
  id_rosto: number;
  last_update: string;
  create_date: string;
  uuid: string;
}